requirements.txt
streamlit
torch
torchvision
fpdf
Pillow
numpy